feat(code-example-insights): refactor sorting logic to use changedLinesCount method and remove redundant code
