+++ app/models/community-course-stage-solution.ts
@@ -13,12 +13,6 @@ import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { memberAction } from 'ember-api-actions';
 import { type FileComparison, FileComparisonFromJSON } from 'codecrafters-frontend/utils/file-comparison';
-import { capitalize } from '@ember/string';
-
-interface EvaluationResult {
-  result: '✅ Passed' | '❌ Failed';
-  check: string;
-}
 
 export default class CommunityCourseStageSolutionModel extends Model.extend(ViewableMixin, VotableMixin) {
   static defaultIncludedResources = ['user', 'language', 'comments', 'comments.user', 'comments.target', 'course-stage'];
@@ -77,52 +71,6 @@ export default class CommunityCourseStageSolutionModel extends Model.extend(View
     });
   }
 
-  get changedLinesCount() {
-    let added = 0,
-      removed = 0;
-
-    for (const changedFile of this.changedFiles) {
-      const diff = changedFile['diff'];
-      const lines = diff.split('\n');
-      added += lines.filter((line: string) => line.startsWith('+')).length;
-      removed += lines.filter((line: string) => line.startsWith('-')).length;
-    }
-
-    return added + removed;
-  }
-
-  get formattedEvaluationResults(): EvaluationResult[] {
-    if (this.evaluations.length === 0) {
-      return [];
-    } else {
-      const results: EvaluationResult[] = [];
-
-      for (const evaluation of this.evaluations) {
-        if (evaluation.result === 'pass') {
-          results.push({
-            result: '✅ Passed',
-            check: evaluation.evaluator.slug,
-          });
-        } else if (evaluation.result === 'fail') {
-          results.push({
-            result: '❌ Failed',
-            check: evaluation.evaluator.slug,
-          });
-        }
-      }
-
-      return results;
-    }
-  }
-
-  get formattedScoreReason() {
-    if (this.isScored && this.scoreReason) {
-      return '⭐' + ' ' + capitalize(this.scoreReason);
-    } else {
-      return 'Not Scored';
-    }
-  }
-
   // We don't render explanations at the moment
   get hasExplanation() {
     return !!this.explanationMarkdown;
@@ -134,10 +82,6 @@ export default class CommunityCourseStageSolutionModel extends Model.extend(View
 
   get isPublishedToPublicGithubRepository() {
     return this.isPublishedToGithub && !this.githubRepositoryIsPrivate;
-  }
-
-  get isScored() {
-    return this.score !== null && this.score > 0;
   }
 
   get screencast() {
